Move client side logger into helpers
