+++ designer/client/src/LinkEdit.tsx
@@ -1,11 +1,11 @@
 import React, { Component } from 'react'
 
+import { logger } from '~/src/common/helpers/logging/logger.js'
 import { SelectConditions } from '~/src/conditions/SelectConditions.jsx'
 import { DataContext } from '~/src/context/DataContext.js'
 import { findPage } from '~/src/data/page/findPage.js'
 import { updateLink } from '~/src/data/page/updateLink.js'
 import { i18n } from '~/src/i18n/i18n.jsx'
-import logger from '~/src/plugins/logger.js'
 
 export class LinkEdit extends Component {
   static contextType = DataContext
