+++ designer/client/src/common/helpers/logging/logger.ts
@@ -1,0 +1,5 @@
+import { pino } from 'pino'
+
+import { loggerOptions } from '~/src/common/helpers/logging/logger-options.js'
+
+export const logger = pino(loggerOptions)
